Un-globalize readRes structure
Now the main function passes the structure readRes directly to the reading thread by (void *) casting, and stores the value of the keys ** array into the readRes structure with the returned value of the thread_join (with necessary char ** casting. readRes is in main but it is no longer a global variable.